+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+# 1.1.0
+- Message history is now scrollable
+- Character limit can now be configured by the host. (defaults to 1000)
+
 # 1.0.0
 Released for v45